// src/infra/model/param-dto/param-dto-model.ts
import { ParamType } from "../../enum/param-type";

/**
 * The DTO model of parameters (data transport model)
 * for all parameters directly from the command line.
 */

export interface ParamDtoEntityModel {
  paramBaseValue: string;
  paramIndex: number;
  paramType: ParamType;
  paramHasValue: boolean;
  paramHasManyValues: boolean;
  paramName: string;
  paramValues: Array<string>;
}

export interface ParamDtoModel {
  params: Array<ParamDtoEntityModel>;
}